refactor: extract region map base terrain tile painting (#3878)
Split sea/land Wang base-tile helpers from render_core into a dedicated
part file to continue Phase 3 flame map modularization.
