+++ app/lib/features/game/flame/region_map/region_map_component.dart
@@ -28,6 +28,7 @@ part 'region_map_component_shared.dart';
 part 'region_map_component_shared_visibility.dart';
 part 'region_map_component_render_orchestrator.dart';
 part 'region_map_component_render_core.dart';
+part 'region_map_component_render_core_base_tiles.dart';
 part 'region_map_component_render_core_overlays.dart';
 part 'region_map_component_render_political_labels.dart';
 part 'region_map_component_render_political_borders.dart';
